refactor : mint-card component
